requirements.txt
# LIMITLESS — dependencias Python (02, 03, 04, 05)
# Instalación: pip install -r requirements.txt

pyyaml>=6.0
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.28.0
